feat: implement GraphHandlerService for handling graph execution and saving; update GraphGateway to use new service; refactor app module and add GraphModule
