+++ packages/backend/src/graph/graph.module.ts
@@ -1,7 +1,8 @@
 import { Module } from '@nestjs/common';
 import { GraphGateway } from './graph.gateway';
+import { GraphService } from 'src/graph.service';
 
 @Module({
-  providers: [GraphGateway],
+  providers: [GraphGateway, GraphService],
 })
 export class GraphModule {}
